requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cffi==1.16.0
cryptography==42.0.3
Django==4.2.10
django-filter==23.5
django-rest-knox==4.2.0
djangorestframework==3.14.0
importlib-metadata==7.0.1
Markdown==3.5.2
pycparser==2.21
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.9.0
zipp==3.17.0
